requirements.txt
Pillow
tkinter

